Fixed Bugs on Navbar, Slider and footer logo.
