+++ projects/container/src/app/components/footer/footer.component.html
@@ -1,7 +1,7 @@
 <footer>
 
   <div class="logo-container">
-    <a routerLink="/" routerLinkActive="active"><img class="logo" src="/../../assets/logo-mrcinema.png" alt="Logo cinema"></a>
+    <a routerLink="/" routerLinkActive="active"><img class="logo" src="/../../assets/img/logo-mrcinema.png" alt="Logo cinema"></a>
     <a routerLink="/" routerLinkActive="active"><h1 class="title">MR.Cinema</h1></a>
   </div>
