+++ projects/container/src/app/components/header/header.component.css
@@ -1,0 +1,144 @@
+@media only screen and (max-width: 768px) {
+  nav,
+  .navigation {
+    flex-direction: column !important;
+    padding-top: 1.5rem !important;
+  }
+  .search,
+  .user {
+    padding: 1.5rem !important;
+  }
+}
+
+nav {
+  width: 100%;
+  background: var(--blue1);
+  background: linear-gradient(175deg, var(--blue1) 30%, var(--blue2) 60%);
+  font-family: 'Raleway', sans-serif;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  gap: 1rem;
+  margin: 0;
+  padding: 2rem 2rem;
+}
+a {
+  text-decoration: none;
+  color: var(--white1);
+}
+ul {
+  list-style: none;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+  padding: 0;
+  margin: 0;
+}
+
+/* Logo */
+.logo-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: start;
+}
+.logo-container a {
+  display: flex;
+}
+.logo-container h1 {
+  color: var(--greenlogo2);
+  font-family:'Rowdies', cursive;
+  font-size: large;
+  margin: 1.2rem 0 0 1.2rem;
+}
+.logo {
+  width: 3rem;
+  height: 2.5rem;
+}
+
+/* Navigation Menu */
+.navigation {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+/* Navigation Menu Animation */
+.navigation li a{
+  /* width: 10%; */
+  position: relative;
+  color: var(--white1);
+  text-decoration: none;
+}
+.navigation li a::before{
+  width: 0;
+  height: 2px;
+  bottom: 0;
+  left: 50%;
+  content: "";
+  position: absolute;
+  background-color: var(--white1);
+  transform: translateX(-50%);
+  transition: width 0.5s ease-in-out;
+}
+.navigation li a:hover::before{
+  width: 100%;
+}
+/* Active Menu Selected*/
+.navigation li a.active {
+  color: var(--white1);
+  font-weight: bold;
+  position: relative;
+}
+.navigation li a.active::before {
+  width: 100%;
+  height: 2px;
+  bottom: 0;
+  left: 0;
+  content: "";
+  position: absolute;
+  background-color: var(--white1);
+  transform: scaleX(0); /* The line begins invisible */
+  transform-origin: left;
+  transition: width 0.5s ease-in-out; /* Added a transition for the animation */
+}
+/* When the 'active' class is applied, the line is animated to appear */
+.navigation li a.active::before {
+  transform: scaleX(1); /* The line becomes visible with an animation */
+}
+
+/* Searchbox */
+.search {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+}
+.search input {
+  height: 2.3rem;
+  border: none;
+  border-top-left-radius: 25px;
+  border-bottom-left-radius: 25px;
+  padding: 0.9rem;
+}
+.search button {
+  height:  2.3rem;
+  border: none;
+  background-color: var(--white1);
+  border-top-right-radius: 25px;
+  border-bottom-right-radius: 25px;
+  padding: 0.5rem 1rem;
+}
+
+/* User Menu */
+.user {
+  display: flex;
+  flex-direction: row;
+}
+.dashboard {
+  background-color: var(--blue1);
+  padding: 0.8rem;
+  border-radius: 10px;
+  font-weight: bold;
+}
