+++ projects/container/src/app/components/header/header.component.html
@@ -1,0 +1,29 @@
+<header>
+  <nav>
+
+    <div class="logo-container">
+      <a routerLink="/">
+        <img class="logo" src="/../../assets/img/logo-mrcinema.png" alt="Logo cinema">
+        <h1>MR.Cinema</h1>
+      </a>
+    </div>
+
+    <ul class="navigation">
+      <li><a routerLink="/movies/list" routerLinkActive="active">Movies</a></li>
+      <li><a routerLink="/movies/colombian-movies" routerLinkActive="active">Colombian Movies</a></li>
+      <li><a routerLink="/candy-shop" routerLinkActive="active">Candy Shop</a></li>
+    </ul>
+
+    <div class="search">
+      <input type="text" placeholder="Search">
+      <button><i-feather name="search"></i-feather></button>
+    </div>
+
+    <ul class="user">
+      <li><i-feather name="user"></i-feather></li>
+      <li><i-feather name="shopping-cart"></i-feather></li>
+      <li><a class="dashboard" routerLink="/dashboard">Dashboard</a></li>
+    </ul>
+
+  </nav>
+<header>
