+++ projects/container/src/app/components/header/header.component.ts
@@ -1,0 +1,25 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-header',
+  templateUrl: './header.component.html',
+  styleUrls: ['./header.component.css'
+  ]
+})
+
+export class HeaderComponent {
+
+/*   menuOptions: string[] = ['Movies', 'Colombian Movies', 'Candy Shop'];
+  selectedOption: string = '';
+
+  selectOption(option: string) {
+    this.selectedOption = option;
+  }
+
+  isOptionSelected(option: string): boolean {
+    return this.selectedOption === option;
+  }
+ */
+
+
+}
