+++ projects/container/src/app/components/sliders/sliders.component.css
@@ -1,70 +1,62 @@
-
 .container-initial{
+  position: relative;
+  width: 100%;
+  height: 16.5rem;
+  background-color: var(--grey1);
   display: flex;
+  flex-direction: column;
   justify-content: center;
   align-items: center;
   overflow: hidden;
-  position: relative;
-  width: 100%;
-  height: 380px;
-  background-color: #DBDFEA;
-  flex-direction: column;
 }
-
 .slider-container {
   position: relative;
-  width: 1535px;
-  height: 380px;
-  background-color: #DBDFEA;
- 
+  width: 80%;
+  height: 100%;
+  background-color: var(--grey1);
+}
+.button-left {
+  position: absolute;
+  background: var(--grey1);
+  border: none;
+  border-radius: 40px;
+  top: 50%;
+  left: 20px;
+  transform: translateY(-50%);
+  padding: 0.9em;
+}
+.button-right {
+  position: absolute;
+  top: 50%;
+  right: 20px;
+  transform: translateY(-50%);
+  background: var(--grey1);
+  border: none;
+  border-radius: 40px;
+  padding: 0.9em;
 }
 .slider-item {
   width: 100%;
   height: 100%;
   object-fit: cover;
 }
-
-.button-left {
-  position: absolute;
-  top: 50%;
-  left: 20px;
-  transform: translateY(-50%);
-  background: #DBDFEA;
-  border: none;
-  border-radius: 40px;
-  padding: 0.9em;
-}
-
-.button-right {
-  position: absolute;
-  top: 50%;
-  right: 20px;
-  transform: translateY(-50%);
-  background: #DBDFEA;
-  border: none;
-  border-radius: 40px;
-  padding: 0.9em;
-}
-
 .pagination {
   position: absolute;
-  bottom: 10px;
+  bottom: 20px;
   left: 0;
   right: 0;
   display: flex;
   justify-content: center;
   align-items: center;
 }
-
 .pagination-dot {
-  width: 10px;
-  height: 10px;
-  background-color: gray;
+  width: 12px;
+  height: 12px;
+  background: var(--grey2);
   border-radius: 50%;
-  margin: 0 5px;
   cursor: pointer;
+  margin: 0 10px;
 }
-
 .pagination-dot.active {
-  background-color: black;
+  background-color: var(--black);
 }
