+++ projects/container/src/app/components/sliders/sliders.component.ts
@@ -1,23 +1,31 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-sliders',
   templateUrl: './sliders.component.html',
   styleUrls: ['./sliders.component.css']
 })
-export class SlidersComponent {
-  images: string[] = [
+export class SlidersComponent implements OnInit {
+
+  public images: string[] = [
+    '../../assets/img/oppenheimer.jpg',
     '../../assets/img/barbie.jpg',
     '../../assets/img/guardianOfGalaxy.jpg',
-    '../../assets/img/oppenheimer.jpg',
     '../../assets/img/spiderman-across.jpg',
   ];
-  currentImage: string;
-  activeIndex: number;
+  public currentImage: string;
+  public activeIndex: number;
 
   constructor() {
     this.currentImage = this.images[0];
     this.activeIndex = 0;
+  }
+  ngOnInit(): void {
+    setInterval(() => {
+      this.activeIndex = this.activeIndex < this.images.length - 1
+       ? this.activeIndex + 1 : 0;
+      this.currentImage = this.images[this.activeIndex];
+    }, 10 * 1000);
   }
 
   previousSlide() {
@@ -34,5 +42,5 @@ export class SlidersComponent {
     this.activeIndex = index;
     this.currentImage = this.images[this.activeIndex];
   }
-}
 
+}
