--- projects/container/src/app/header/header.component.css
@@ -1,98 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Raleway&family=Rowdies&display=swap');
-
-    nav{
-        background: var(--blued);
-        background: linear-gradient(175deg, var(--blued) 30%, var(--bluel) 60%);
-	    display: flex;
-	    justify-content: center;
-        width: 100%;
-        gap: 5em;
-	    flex-direction: row;
-		margin: 0;
-		padding: 1em 4.5em;
-        font-family: 'Raleway', sans-serif;
-	}
-    a{
-        text-decoration: none;
-        color: #ffffff;
-        
-    }
-    h1{
-    justify-content: center;
-    color: #ffffff;
-    font-family:'Rowdies', cursive;
-    }
-    .logo-container{
-        display: flex;
-	    justify-content: center;
-	    flex-direction: row;
-        align-items: center;
-    }
-    .logo-container a {
-        display: flex;
-    }
-    ul{
-        list-style: none;
-        display: flex;
-	    gap: 2em;
-        align-items: center;
-        padding: 0;
-        margin: 0;  
-    }
-    input{
-        border: none;
-        border-top-left-radius:25px ;
-        border-bottom-left-radius: 25px;
-        padding: 0.9em ;
-        height:  3.5em;
-    }
-    button{
-        border: none;
-        background-color: #ffffff;
-        border-top-right-radius:25px ;
-        border-bottom-right-radius: 25px;
-        padding: 0.5em;
-        padding-right: 1.5em;
-        height:  3.5em;
-    }
-    .search{
-        display: flex;
-        justify-content: center;
-        flex-direction: row;
-        align-items: center;
-    }
-    .dashboard{
-        background-color: var(--blued);
-        padding: 0.7em;
-        border-radius: 10px;
-        font-weight: bold;
-    }
-    .logo{ 
-        width: 4.375em;
-        height: 3.438em;
-    }
-    .navigation li a{
-        position: relative;
-        display: inline-block;
-        padding-bottom: 10px;
-        text-decoration: none;
-        color: #ffffff;
-    }
-
-    .navigation li a::before{
-        content: "";
-        position: absolute;
-        bottom: 0;
-        left: 50%;
-        width: 0;
-        height: 2px;
-        background-color: #ffffff;
-        transform: translateX(-50%);
-        transition: width 0.5s ease-in-out;
-       
-    }
-
-    .navigation li a:hover::before{
-        width: 100%;
-    }
-
