--- projects/container/src/app/header/header.component.html
@@ -1,37 +1,0 @@
-<header>
-
-    <nav>
-
-        <div class="logo-container">
-            <a routerLink="/">
-                <img  class="logo" src="/../../assets/logo-mrcinema.png" alt="Logo cinema">
-                <span><h1>{{title}}</h1></span>
-            </a>
-        </div>
-
-        <ul class="navigation">
-            <li><a routerLink="/movies/list">Movies</a></li>
-            <li><a routerLink="/movies/colombian-movies">Colombian Movies</a></li>
-            <li><a routerLink="/candy-shop">Candy Shop</a></li>
-        </ul>
-
-        <div class="search">
-            <input type="text" placeholder="Search">
-            <button><i-feather name="search"></i-feather></button>
-        </div>
-
-        <ul class="user">
-            <li>
-                <i-feather name="user"></i-feather>
-            </li>
-            <li>
-                <i-feather name="shopping-cart"></i-feather>
-            </li>
-            <li>
-                <a class="dashboard" routerLink="/dashboard">Dashboard</a>
-            </li>
-        </ul>
-
-    </nav>
-
-<header>
