--- projects/container/src/app/header/header.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.css'
-  ]
-})
-
-export class HeaderComponent {
-
-  public title = 'Mr.Cinema';
-}
